third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/live-edit-breakpoints-expected.txt

Tests breakpoints are correctly dimmed and restored in JavaScriptSourceFrame during live edit.


Running: testEditUndo
Setting breakpoint:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Editing source:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: VMXX edit-me-breakpoints.js, lineNumber: 2, project type: debugger
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Undoing source editing:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Finally removing breakpoint:
    Dumping breakpoint storage
    Dumping breakpoint locations
    Dumping breakpoints

Running: testEditCommit
Setting breakpoint:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Editing source:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: VMXX edit-me-breakpoints.js, lineNumber: 2, project type: debugger
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Committing edited source:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Finally removing breakpoint:
    Dumping breakpoint storage
    Dumping breakpoint locations
    Dumping breakpoints

Running: testEditCommitFailEditCommit
Setting breakpoint:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Editing source:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: VMXX edit-me-breakpoints.js, lineNumber: 2, project type: debugger
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Committing edited source:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: VMXX edit-me-breakpoints.js, lineNumber: 2, project type: debugger
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Editing source again so that live edit could succeed:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: VMXX edit-me-breakpoints.js, lineNumber: 2, project type: debugger
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Committing edited source again:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Finally removing breakpoint:
    Dumping breakpoint storage
    Dumping breakpoint locations
    Dumping breakpoints

Running: testEditCommitFailUndoCommit
Setting breakpoint:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Editing source:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: VMXX edit-me-breakpoints.js, lineNumber: 2, project type: debugger
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Committing edited source:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: VMXX edit-me-breakpoints.js, lineNumber: 2, project type: debugger
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Undoing source editing:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: VMXX edit-me-breakpoints.js, lineNumber: 2, project type: debugger
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Committing edited source again:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
    Dumping breakpoints
        url: edit-me-breakpoints.js, lineNumber: 2, project type: network
Finally removing breakpoint:
    Dumping breakpoint storage
    Dumping breakpoint locations
    Dumping breakpoints
